OpenMP Task: Drop 'inout' depend and use more iterators when possible
Drop the 'inout' in depend clause for tasks so that all 'in' variables are
pinned correctly when depending on another parititon.

Use iterators if needed in the 'getDependenciesAll' method so that in
reductions and affectation only jobs there is less to write in the depend
clause and push more informations to the OpenMP runtime.