OMP Task & CPPLIB: Print a 'XXX' message when generated buggy tasks
Some times, we generate tasks with the same outputs, thus we have:
Task1: A => B
Task2: C => B
Task3: B => D

It seems unclear when the tasks 3 must begin. The way the programs are
written the task 3 must begin after the 2 B are produced, but it's seems
that OpenMP doesn't do that <= FIXME (in the generation/CPPLIB)!