REFACTOR: Merge the 'CartesianPartition2D' and 'CartesianMesh2D'.
This will permit to create 'partitions' in the Mesh2D object and make OpenMP
pin them to have all the memory that needs to be synchronized. This change is
destructive with other generators, i.e. they might break, rename the new
CartesianMesh2D into another class name to fix that (don't do it for the moment).